fix: products

+++ src/domains/product.domain.ts
@@ -1,26 +1,37 @@
 import { ProductRepository } from '../repositories/product.repo';
-import { Product, CreateProductRequest, UpdateProductRequest } from '../types/product.type';
+import { ProductsCategoryRepository } from '../repositories/products_category.repo';
 import { wrapperData } from '../utils/wrapper';
-import { DataNotFound } from '../utils/errors';
+import { BadRequest, DataNotFound } from '../utils/errors';
 import { WrapperData, PaginationMeta } from '../types/wrapper.type';
+import {
+  Product,
+  CreateProductRequest,
+  UpdateProductRequest,
+  ListProductRequest
+} from '../types/product.type';
 
 export class ProductDomain {
   private repo: ProductRepository;
+  private category: ProductsCategoryRepository;
 
   constructor() {
     this.repo = new ProductRepository();
+    this.category = new ProductsCategoryRepository()
   }
 
-  async getAllProducts(limit: number = 50, offset: number = 0): Promise<{ data: Product[], meta: PaginationMeta }> {
+  async getAllProducts(params: ListProductRequest): Promise<{ data: Product[], meta: PaginationMeta }> {
+    const limit = parseInt(params.size)
+    const offset = (parseInt(params.page) - 1) * limit
+
     const result = await this.repo.getAllProducts(limit, offset);
-    const totalPages = Math.ceil(result.total / limit);
-    const currentPage = Math.floor(offset / limit) + 1;
+
     const meta: PaginationMeta = {
       total: result.total,
       limit,
-      totalPages,
-      currentPage
+      totalPages: result.total > 0 ? Math.ceil(result.total / limit) : 1,
+      currentPage: Math.floor(offset / limit) + 1
     };
+
     return { data: result.data, meta };
   }
 
@@ -32,35 +43,55 @@ export class ProductDomain {
     return wrapperData(data, null)
   }
 
-  async createProduct(productData: CreateProductRequest): Promise<Product> {
-    this.validateCreateProduct(productData);
-    return this.repo.createProduct(productData);
+  async createProduct(productData: CreateProductRequest): Promise<WrapperData> {
+    // check name
+    const checkName = await this.repo.getProductByName(productData.name.toLowerCase())
+    if (checkName) {
+      return wrapperData(null, BadRequest('Nama produk sudah digunakan'))
+    }
+
+    // check category
+    const checkCategory = await this.category.getProductsCategoryById(productData.categoryId)
+    if (!checkCategory) {
+      return wrapperData(null, DataNotFound('Kategori tidak ditemukan'))
+    }
+
+    const result = await this.repo.createProduct(productData);
+
+    return wrapperData(result, null)
   }
 
-  async updateProduct(id: string, updates: UpdateProductRequest): Promise<Product | null> {
-    this.validateUpdateProduct(updates);
-    return this.repo.updateProduct(id, updates);
+  async updateProduct(id: string, updates: UpdateProductRequest): Promise<WrapperData> {
+    // check data id
+    const checkId = await this.repo.getProductById(id)
+    if (!checkId) {
+      return wrapperData(null, DataNotFound())
+    }
+
+    // check name
+    const checkName = await this.repo.getProductByName(updates.name)
+    if (checkName && checkName.id !== id) {
+      return wrapperData(null, BadRequest())
+    }
+
+    // check category
+    const checkCategory = await this.category.getProductsCategoryById(updates.categoryId)
+    if (!checkCategory) {
+      return wrapperData(null, DataNotFound('Kategori tidak ditemukan'))
+    }
+
+    const result = await this.repo.updateProduct(id, updates);
+    return wrapperData(result, null)
   }
 
-  async deleteProduct(id: string): Promise<boolean> {
-    return this.repo.deleteProduct(id);
-  }
+  async deleteProduct(id: string): Promise<WrapperData> {
+    // check data id
+    const checkId = await this.repo.getProductById(id)
+    if (!checkId) {
+      return wrapperData(null, DataNotFound())
+    }
 
-  private validateCreateProduct(product: CreateProductRequest): void {
-    if (!product.name || product.name.trim().length === 0) {
-      throw new Error('Product name is required');
-    }
-    if (product.price <= 0) {
-      throw new Error('Product price must be greater than 0');
-    }
-  }
-
-  private validateUpdateProduct(updates: UpdateProductRequest): void {
-    if (updates.name !== undefined && (!updates.name || updates.name.trim().length === 0)) {
-      throw new Error('Product name cannot be empty');
-    }
-    if (updates.price !== undefined && updates.price <= 0) {
-      throw new Error('Product price must be greater than 0');
-    }
+    const result = await this.repo.deleteProduct(id);
+    return wrapperData({ code: result?.code }, null)
   }
 }
